requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
jinja2==3.1.2
apscheduler==3.10.4
lxml==4.9.3
python-slugify==8.0.1
gitpython==3.1.40
anthropic>=0.39.0
requests==2.31.0
